// src/app/shared/enum/enumerator.enum.ts
export enum Enumerator {
    TOASTER_DURATION_TIME = 5000,
    HTTP_RESPONSE_OK = 200,
    HTTP_RESPONSE_OK_STORAGE = 201,
    HTTP_RESPONSE_UNAUTHORIZED = 401,
    HTTP_RESPONSE_BADREQUEST = 400,
    HTTP_RESPONSE_SERVERERROR = 500,
    HTTP_RESPONSE_UNKNOWN = 0,
    HEADING_GROUP_ONE = 1,
    HEADING_GROUP_TWO = 2,
    HEADING_GROUP_THREE = 3,
    INFRASTRUCTURETYPE_RESERVOIR_TANK = 38,
    INFRASTRUCTURETYPE_CORRAL_BARN = 39,
    INFRASTRUCTURETYPE_CATALOG = 14,
    INFRASTRUCTURE_EXECUTED_COST = 15,
    MIN_PHOTOS_INFRASTRUCTURETYPE_RESERVOIR_TANK = 16,
    MAX_PHOTOS_INFRASTRUCTURETYPE_RESERVOIR_TANK = 17,
    MIN_PHOTOS_INFRASTRUCTURETYPE_CORRAL_BARN = 18,
    MAX_PHOTOS_INFRASTRUCTURETYPE_CORRAL_BARN = 19,
    MIN_SIZE_PHOTOS_INFRASTRUCTURE = 20,
    MAX_SIZE_PHOTOS_INFRASTRUCTURE = 21,
    ID_ITEM_TRACTORS = 4,
    ID_ITEM_WORKING_CAPITAL = 5,
    ID_ACTIVIDAD = 13,
    ID_SUPPORTS_INVESTMENT_PHOTO = 36,
    ID_SUPPORTS_BILLS_PHOTO = 37,
    MAX_ACTIVIDAD_POR_TIPO = 23,
    DIAN_PURCHASE_INVOICE = 33,
    ID_MACHINERY = 31,
    ID_MACHINERY_PLATE = 32,
    PHOTOGRAPH_FINANCED_PROJECT = 34,
    RUBRO_TYPE = 22,
    MIN_PHOTOS_MACHINERY_MACHINERY_ACQUIRED = 30,
    MAX_PHOTOS_MACHINERY_MACHINERY_ACQUIRED = 29,
    MIN_PHOTOS_MACHINERY_MACHINERY_PLATE = 32,
    MAX_PHOTOS_MACHINERY_MACHINERY_PLATE = 31,
    MIN_PHOTOS_MACHINERY_DIAN_INVOICE = 34,
    MAX_PHOTOS_MACHINERY_DIAN_INVOICE = 33,
    MIN_SIZE_PHOTOS_MACHINERY = 38,
    MAX_SIZE_PHOTOS_MACHINERY = 37,
    MIN_PHOTOS_CAPITAL_FINANCED_PROJECT = 26,
    MAX_PHOTOS_CAPITAL_FINANCED_PROJECT = 25,
    MIN_PHOTOS_CAPITAL_INVESTMENT_SUPPORT = 28,
    MAX_PHOTOS_CAPITAL_INVESTMENT_SUPPORT = 27,
    MAX_SIZE_PHOTOS_CAPITAL = 35,
    MIN_SIZE_PHOTOS_CAPITAL = 36,
    MACHINERY_EXECUTED_COST = 40,
    CAPITAL_EXECUTED_COST = 39,
    CODIGO_CAPITAL_TRABAJO = 68,
    CODIGO_INFRASTRUCTURA = 106,
    CODIGO_MAQUINARIA_EQUIPOS = 87,
    INTENTOS_FALLIDOS_INGRESO_TOKEN = 3,
    CAMERA_PERMISSION_ERROR = 20

}